:recycle: refactor house rooms / view into own files
