+++ house/bedroom/bathroom.lua
@@ -1,0 +1,15 @@
+require "house.room"
+
+local Bathroom =
+    Child(
+    "bathroom",
+    "n",
+    2,
+    2,
+    {
+        type = "door",
+        locked = true
+    }
+)
+
+return Bathroom
