+++ house/bedroom/closet.lua
@@ -1,0 +1,13 @@
+local Closet =
+    Child(
+    "closet",
+    "s",
+    1,
+    1,
+    {
+        type = "door",
+        locked = true
+    }
+)
+
+return Closet
